requirements.txt
#
openai==1.68.2
python-dotenv==1.0.1
requests==2.31.0